MyApps-7: Improve menu entries.
